Move all of css into styles folder, Add card under Landing page
