--- components/Cards.js
@@ -1,12 +1,0 @@
-import './Cards.scss';
-
-const Cards = () => (
-
-  <div className="container-card">
-    <img className="cards" src="/smallimage.jpg" alt="card" />
-    <img className="cards" src="/smallimage.jpg" alt="card" />
-    <img className="cards" src="/smallimage.jpg" alt="card" />
-  </div>
-);
-
-export default Cards;
