--- components/Cards.scss
@@ -1,13 +1,0 @@
-@import "../theme/theme";
-
-.container-card{
-    display: flex;
-    flex-direction: row;
-    flex: nowrap;
-}
-
-.cards{
-
-    width: 100%;
-
-}
