+++ components/Layout.js
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import Header from './Header';
 import NavBar from './NavBar';
 
-import './Layout.scss';
+import '../styles/Layout.scss';
 
 /* inorder to have a document title on every page, use the <head> block
 in the Layout component template be sure to import Head from 'next/head' to modify the title. */
