+++ pages/index.js
@@ -1,7 +1,7 @@
 // import Layout from '../components/Layout';
 import Feature from '../components/Feature';
 import Slider from '../components/Slider';
-import '../components/index.scss';
+import '../styles/index.scss';
 
 const Index = () => (
   <div className="main-container">
