+++ styles/Feature.scss
@@ -15,7 +15,7 @@
     height: 100%;
 }
 
-img{
+.hero-img{
     position: absolute;
     top: 0;
     left: 0;
