+++ styles/NavBar.scss
