+++ styles/Slider.scss
