+++ styles/Swiper.scss
